Break out the single WebP Uploads test class into multiple test classes, exactly following the main file breakdown.
