+++ tests/modules/images/webp-uploads/helper-tests.php
@@ -1,0 +1,236 @@
+<?php
+/**
+ * Tests for webp-uploads module helper.php.
+ *
+ * @package performance-lab
+ * @group   webp-uploads
+ */
+
+use PerformanceLab\Tests\TestCase\ImagesTestCase;
+
+class WebP_Uploads_Helper_Tests extends ImagesTestCase {
+
+	/**
+	 * Return an error when creating an additional image source with invalid parameters
+	 *
+	 * @dataProvider data_provider_invalid_arguments_for_webp_uploads_generate_additional_image_source
+	 *
+	 * @test
+	 */
+	public function it_should_return_an_error_when_creating_an_additional_image_source_with_invalid_parameters( $attachment_id, $size_data, $mime, $destination_file = null ) {
+		$this->assertInstanceOf( WP_Error::class, webp_uploads_generate_additional_image_source( $attachment_id, $size_data, $mime, $destination_file ) );
+	}
+
+	public function data_provider_invalid_arguments_for_webp_uploads_generate_additional_image_source() {
+		yield 'when trying to use an attachment ID that does not exists' => array(
+			PHP_INT_MAX,
+			array(),
+			'image/webp',
+		);
+
+		add_filter( 'wp_image_editors', '__return_empty_array' );
+		yield 'when no editor is present' => array(
+			$this->factory->attachment->create_upload_object( TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/car.jpeg' ),
+			array(),
+			'image/avif',
+		);
+
+		remove_filter( 'wp_image_editors', '__return_empty_array' );
+		yield 'when using a mime that is not supported' => array(
+			$this->factory->attachment->create_upload_object( TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/car.jpeg' ),
+			array(),
+			'image/avif',
+		);
+
+		yield 'when no dimension is provided' => array(
+			$this->factory->attachment->create_upload_object( TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/car.jpeg' ),
+			array(),
+			'image/webp',
+		);
+
+		yield 'when both dimensions are negative numbers' => array(
+			$this->factory->attachment->create_upload_object( TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/car.jpeg' ),
+			array(
+				'width'  => -10,
+				'height' => -20,
+			),
+			'image/webp',
+		);
+
+		yield 'when both dimensions are zero' => array(
+			$this->factory->attachment->create_upload_object( TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/car.jpeg' ),
+			array(
+				'width'  => 0,
+				'height' => 0,
+			),
+			'image/webp',
+		);
+	}
+
+	/**
+	 * Create an image with the default suffix in the same location when no destination is specified
+	 *
+	 * @test
+	 */
+	public function it_should_create_an_image_with_the_default_suffix_in_the_same_location_when_no_destination_is_specified() {
+		$attachment_id = $this->factory->attachment->create_upload_object( TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/car.jpeg' );
+		$size_data     = array(
+			'width'  => 300,
+			'height' => 300,
+			'crop'   => true,
+		);
+
+		$result    = webp_uploads_generate_additional_image_source( $attachment_id, $size_data, 'image/webp' );
+		$file      = get_attached_file( $attachment_id );
+		$directory = trailingslashit( pathinfo( $file, PATHINFO_DIRNAME ) );
+		$name      = pathinfo( $file, PATHINFO_FILENAME );
+
+		$this->assertIsArray( $result );
+		$this->assertArrayHasKey( 'filesize', $result );
+		$this->assertArrayHasKey( 'file', $result );
+		$this->assertStringEndsWith( '300x300.webp', $result['file'] );
+		$this->assertFileExists( "{$directory}{$name}-300x300.webp" );
+	}
+
+	/**
+	 * Create a file in the specified location with the specified name
+	 *
+	 * @test
+	 */
+	public function it_should_create_a_file_in_the_specified_location_with_the_specified_name() {
+		$attachment_id = $this->factory->attachment->create_upload_object( TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/car.jpeg' );
+		$size_data     = array(
+			'width'  => 300,
+			'height' => 300,
+			'crop'   => true,
+		);
+
+		$result = webp_uploads_generate_additional_image_source( $attachment_id, $size_data, 'image/webp', '/tmp/image.jpg' );
+
+		$this->assertIsArray( $result );
+		$this->assertArrayHasKey( 'filesize', $result );
+		$this->assertArrayHasKey( 'file', $result );
+		$this->assertStringEndsWith( 'image.webp', $result['file'] );
+		$this->assertFileExists( '/tmp/image.webp' );
+	}
+
+	/**
+	 * Prevent processing an image with corrupted metadata
+	 *
+	 * @dataProvider provider_with_modified_metadata
+	 *
+	 * @test
+	 */
+	public function it_should_prevent_processing_an_image_with_corrupted_metadata( callable $callback, $size ) {
+		$attachment_id = $this->factory->attachment->create_upload_object(
+			TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/balloons.webp'
+		);
+		$metadata      = wp_get_attachment_metadata( $attachment_id );
+		wp_update_attachment_metadata( $attachment_id, $callback( $metadata ) );
+		$result = webp_uploads_generate_image_size( $attachment_id, $size, 'image/webp' );
+
+		$this->assertTrue( is_wp_error( $result ) );
+		$this->assertSame( 'image_mime_type_invalid_metadata', $result->get_error_code() );
+	}
+
+	public function provider_with_modified_metadata() {
+		yield 'using a size that does not exists' => array(
+			function ( $metadata ) {
+				return $metadata;
+			},
+			'not-existing-size',
+		);
+
+		yield 'removing an existing metadata simulating that the image size still does not exists' => array(
+			function ( $metadata ) {
+				unset( $metadata['sizes']['medium'] );
+
+				return $metadata;
+			},
+			'medium',
+		);
+
+		yield 'when the specified size is not a valid array' => array(
+			function ( $metadata ) {
+				$metadata['sizes']['medium'] = null;
+
+				return $metadata;
+			},
+			'medium',
+		);
+	}
+
+	/**
+	 * Prevent to create an image size when attached file does not exists
+	 *
+	 * @test
+	 */
+	public function it_should_prevent_to_create_an_image_size_when_attached_file_does_not_exists() {
+		$attachment_id = $this->factory->attachment->create_upload_object(
+			TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/leafs.jpg'
+		);
+		$file          = get_attached_file( $attachment_id );
+
+		$this->assertFileExists( $file );
+		wp_delete_file( $file );
+		$this->assertFileDoesNotExist( $file );
+
+		$result = webp_uploads_generate_image_size( $attachment_id, 'medium', 'image/webp' );
+		$this->assertTrue( is_wp_error( $result ) );
+		$this->assertSame( 'original_image_file_not_found', $result->get_error_code() );
+	}
+
+	/**
+	 * Prevent to create a subsize if the image editor does not exists
+	 *
+	 * @test
+	 */
+	public function it_should_prevent_to_create_a_subsize_if_the_image_editor_does_not_exists() {
+		$attachment_id = $this->factory->attachment->create_upload_object(
+			TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/leafs.jpg'
+		);
+
+		// Make sure no editor is available.
+		add_filter( 'wp_image_editors', '__return_empty_array' );
+		$result = webp_uploads_generate_image_size( $attachment_id, 'medium', 'image/webp' );
+		$this->assertTrue( is_wp_error( $result ) );
+		$this->assertSame( 'image_mime_type_not_supported', $result->get_error_code() );
+	}
+
+	/**
+	 * Prevent to upload a mime that is not supported by WordPress
+	 *
+	 * @test
+	 */
+	public function it_should_prevent_to_upload_a_mime_that_is_not_supported_by_wordpress() {
+		$attachment_id = $this->factory->attachment->create_upload_object(
+			TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/leafs.jpg'
+		);
+		$result        = webp_uploads_generate_image_size( $attachment_id, 'medium', 'image/avif' );
+		$this->assertTrue( is_wp_error( $result ) );
+		$this->assertSame( 'image_mime_type_invalid', $result->get_error_code() );
+	}
+
+	/**
+	 * Prevent to process an image when the editor does not support the format
+	 *
+	 * @test
+	 */
+	public function it_should_prevent_to_process_an_image_when_the_editor_does_not_support_the_format() {
+		// Make sure no editor is available.
+		$attachment_id = $this->factory->attachment->create_upload_object(
+			TESTS_PLUGIN_DIR . '/tests/testdata/modules/images/leafs.jpg'
+		);
+
+		add_filter(
+			'wp_image_editors',
+			function () {
+				return array( 'WP_Image_Doesnt_Support_WebP' );
+			}
+		);
+
+		$result = webp_uploads_generate_image_size( $attachment_id, 'medium', 'image/webp' );
+		$this->assertTrue( is_wp_error( $result ) );
+		$this->assertSame( 'image_mime_type_not_supported', $result->get_error_code() );
+	}
+}
